Rutas protegidas

+++ src/routes/generos.js
@@ -4,24 +4,28 @@ const { isUrl, nombreCapitalizado } = require("../helpers/verificadores");
 ////////////////////////////////////////////////////////////////////
 server.get("/", async (req, res) => {
   try {
-    const { name } = req.query;
-    if (name) {
-      const nombre = nombreCapitalizado(name);
-      const genero = await Genero.findOne({
-        where: { nombre: nombre },
-        include: [{ model: Peliculaserie, as: "peliculaseries" }],
-      });
-      if (genero) {
-        res.json(genero);
+    if (req.user) {
+      const { name } = req.query;
+      if (name) {
+        const nombre = nombreCapitalizado(name);
+        const genero = await Genero.findOne({
+          where: { nombre: nombre },
+          include: [{ model: Peliculaserie, as: "peliculaseries" }],
+        });
+        if (genero) {
+          res.json(genero);
+        } else {
+          res.sendStatus(404);
+        }
       } else {
-        res.sendStatus(404);
+        const allGenres = await Genero.findAll({
+          include: [{ model: Peliculaserie, as: "peliculaseries" }],
+        });
+
+        return res.json(allGenres); //retorna array vacio si no hay generos
       }
     } else {
-      const allGenres = await Genero.findAll({
-        include: [{ model: Peliculaserie, as: "peliculaseries" }],
-      });
-
-      return res.json(allGenres); //retorna array vacio si no hay generos
+      res.sendStatus(401);
     }
   } catch {
     return res.sendStatus(500);
@@ -31,17 +35,22 @@ server.get("/", async (req, res) => {
 ////////////////////////////////////////////////////////////////CRUD
 server.post("/", async (req, res) => {
   const { imagen, nombre } = req.body;
-  try {
-    if (imagen && isUrl(imagen) && nombre) {
-      const nombrePropio = nombreCapitalizado(nombre);
 
-      const genero = await Genero.create({
-        imagen: imagen,
-        nombre: nombrePropio,
-      });
-      return res.sendStatus(201);
+  try {
+    if (req.user) {
+      if (imagen && isUrl(imagen) && nombre) {
+        const nombrePropio = nombreCapitalizado(nombre);
+
+        const genero = await Genero.create({
+          imagen: imagen,
+          nombre: nombrePropio,
+        });
+        return res.sendStatus(201);
+      } else {
+        return res.sendStatus(400);
+      }
     } else {
-      return res.sendStatus(400);
+      res.sendStatus(401);
     }
   } catch {
     return res.sendStatus(500);
@@ -51,11 +60,15 @@ server.post("/", async (req, res) => {
 server.get("/:id", async (req, res) => {
   const { id } = req.params;
   try {
-    const genero = await Genero.findOne({ where: { id: id } });
-    if (genero) {
-      return res.sendStatus(200);
+    if (req.user) {
+      const genero = await Genero.findOne({ where: { id: id } });
+      if (genero) {
+        return res.sendStatus(200);
+      } else {
+        return res.sendStatus(404);
+      }
     } else {
-      return res.sendStatus(404);
+      res.sendStatus(401);
     }
   } catch {
     return res.sendStatus(500);
@@ -67,21 +80,25 @@ server.put("/:id", async (req, res) => {
   const { imagen, nombre } = req.body;
 
   try {
-    const existe = await Genero.findOne({ where: { id: id } });
-    if (existe) {
-      let obj = {}; //usamos un objeto para ahorrar condicionales
-      if (imagen && isUrl(imagen)) obj = { ...obj, imagen: imagen }; //verificamos que imagen sea una url
-      if (nombre) {
-        nombre = nombreCapitalizado(nombre); //capitalizamos nombre si son dos o mas palabras se capitaliza la primera letra
-        obj = { ...obj, nombre: nombre };
+    if (req.user) {
+      const existe = await Genero.findOne({ where: { id: id } });
+      if (existe) {
+        let obj = {}; //usamos un objeto para ahorrar condicionales
+        if (imagen && isUrl(imagen)) obj = { ...obj, imagen: imagen }; //verificamos que imagen sea una url
+        if (nombre) {
+          nombre = nombreCapitalizado(nombre); //capitalizamos nombre si son dos o mas palabras se capitaliza la primera letra
+          obj = { ...obj, nombre: nombre };
+        }
+        const actualizado = await Genero.update(
+          { imagen: obj.imgen && obj.imagen, nombre: obj.nombre && obj.nombre },
+          { where: { id: id } }
+        );
+        return res.sendStatus(200);
+      } else {
+        return res.sendStatus(404);
       }
-      const actualizado = await Genero.update(
-        { imagen: obj.imgen && obj.imagen, nombre: obj.nombre && obj.nombre },
-        { where: { id: id } }
-      );
-      return res.sendStatus(200);
     } else {
-      return res.sendStatus(404);
+      res.sendStatus(401);
     }
   } catch {
     return res.sendStatus(500);
@@ -91,12 +108,16 @@ server.put("/:id", async (req, res) => {
 server.delete("/:id", async (req, res) => {
   const { id } = req.params;
   try {
-    const existe = await Genero.findOne({ where: { id: id } });
-    if (existe) {
-      const eliminado = await Genero.destroy({ where: { id: id } });
-      res.sendStatus(200);
+    if (req.user) {
+      const existe = await Genero.findOne({ where: { id: id } });
+      if (existe) {
+        const eliminado = await Genero.destroy({ where: { id: id } });
+        res.sendStatus(200);
+      } else {
+        res.sendStatus(404);
+      }
     } else {
-      res.sendStatus(404);
+      res.sendStatus(401);
     }
   } catch {
     res.sendStatus(500);
